+++ _examples/routing/mvc/controllers/user.go
@@ -42,7 +42,7 @@ func (c *User) Get() {
 }
 
 /* Can use more than one, the factory will make sure
-that the correct http methods are being registed  for this
+that the correct http methods are being registered  for this
 controller, uncommend these if you want:
 
 func (c *User) Post() {}