+++ _examples/sessions/standalone/main.go
@@ -71,7 +71,7 @@ func main() {
 
 	app.Get("/update", func(ctx context.Context) {
 		// updates expire date
-		sess.ShiftExpiraton(ctx)
+		sess.ShiftExpiration(ctx)
 	})
 
 	app.Get("/destroy", func(ctx context.Context) {