+++ iris.go
@@ -32,7 +32,7 @@ import (
 
 const (
 	// Version is the current version number of the Iris Web Framework.
-	Version = "8.2.5"
+	Version = "8.2.6"
 )
 
 // HTTP status codes as registered with IANA.